test: 리팩토링 된 LottoMachine 클래스에 맞게 테스트 수정
