--- __tests__/LottoGeneratorTest.js
@@ -1,34 +1,0 @@
-const LottoGenerator = require("../src/domain/LottoGenerator");
-const MissionUtils = require("@woowacourse/mission-utils");
-
-const mockRandoms = (numbers) => {
-  MissionUtils.Random.pickUniqueNumbersInRange = jest.fn();
-  numbers.reduce((acc, number) => {
-    return acc.mockReturnValueOnce(number);
-  }, MissionUtils.Random.pickUniqueNumbersInRange);
-};
-
-describe("로또 생성기 클래스 테스트", () => {
-  test("로또 생성기 클래스를 생성할 때 입력한 금액만큼 로또를 생성한다.", () => {
-    const mockArray = [
-      [8, 21, 23, 41, 42, 43],
-      [3, 5, 11, 16, 32, 38],
-      [7, 11, 16, 35, 36, 44],
-      [1, 8, 11, 31, 41, 42],
-      [13, 14, 16, 38, 42, 45],
-      [7, 11, 30, 40, 42, 43],
-      [2, 13, 22, 32, 38, 45],
-      [1, 3, 5, 14, 22, 45],
-    ];
-    mockRandoms(mockArray);
-
-    const lottoGenerator = new LottoGenerator();
-    const lottos = lottoGenerator.getLottos(8);
-
-    lottos.forEach((lotto, idx) => {
-      expect(lotto.numbers).toEqual(mockArray[idx]);
-    });
-
-    MissionUtils.Console.close();
-  });
-});
